[refactor] rename common/protocols to common/protocol_client
