+++ examples/BUILD.bazel
@@ -10,6 +10,6 @@ cc_binary(
   deps = [
     "@com_github_grpc_grpc//:grpc++",
     "@com_google_protobuf//:protobuf_lite",
-    "//src/common/protocols:master_metadata_service_client",
+    "//src/common/protocol_client:master_metadata_service_client",
   ],
 )
