+++ src/common/protocol_client/master_metadata_service_client.h
@@ -1,5 +1,5 @@
-#ifndef GFS_COMMON_PROTOCOLS_MASTER_SERVICE_CLIENT_H_
-#define GFS_COMMON_PROTOCOLS_MASTER_SERVICE_CLIENT_H_
+#ifndef GFS_COMMON_PROTOCOL_CLIENT_MASTER_SERVICE_CLIENT_H_
+#define GFS_COMMON_PROTOCOL_CLIENT_MASTER_SERVICE_CLIENT_H_
 
 #include <memory>
 
@@ -88,4 +88,4 @@ class MasterMetadataServiceClient {
 }  // namespace service
 }  // namespace gfs
 
-#endif  // GFS_COMMON_PROTOCOLS_MASTER_SERVICE_CLIENT_H_
+#endif  // GFS_COMMON_PROTOCOL_CLIENT_MASTER_SERVICE_CLIENT_H_
